requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anthropic==0.18.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiosqlite==0.19.0
httpx==0.26.0
authlib==1.3.0
itsdangerous==2.1.2
jinja2==3.1.3
websockets==12.0
